+++ .github/release.config.js
@@ -22,10 +22,10 @@ module.exports = {
             "@octorelease/changelog",
             {
                 displayNames: {
-                    "cli": "IBM CICS Plug-in for Zowe CLI",
-                    "sdk": "IBM CICS for Zowe SDK",
-                    "vsce": "IBM CICS Extension for Zowe Explorer",
-                    "vsce-api": "IBM CICS Extension for Zowe Explorer API",
+                    "cli": "Zowe CLI Plug-in for IBM CICS Transaction Server",
+                    "sdk": "Zowe SDK for IBM CICS Transaction Server",
+                    "vsce": "Zowe Explorer for IBM CICS Transaction Server",
+                    "vsce-api": "Zowe Explorer for IBM CICS Transaction Server API",
                 },
                 headerLine: "## Recent Changes",
             },